refactor: components/common/styled 폴더구조 변경
