+++ src/components/common/index.ts
@@ -1,8 +1,9 @@
 export { default as AlertLabel } from "./AlertLabel";
-export { Button } from "./Button";
 export { default as CommonLayout } from "./CommonLayout";
-export { Input } from "./Input";
 export { default as MainHeader } from "./MainHeader";
 export { default as MainLayout } from "./MainLayout";
 export { default as NavHeader } from "./NavHeader";
 export { default as NavWrapper } from "./NavWrapper";
+export { Button } from "./styled/Button";
+export { Input } from "./styled/Input";
+export { LabelHidden } from "./styled/LabelHidden";
